feat: Introduce creature type selection
Add the ability to switch between different creature types (MOSASAUR and KAIJU). This includes updating the `CreatureType` enum, passing the selected type to the `GameCanvas` component, and adjusting the prey bite radius based on the creature type. A new "Mutate" button is added to the `Controls` component to facilitate this change.